Added leoni, loreal, maersk, mahle, mairon, marquardt and marriot scrapers
